联动删除

+++ database/migrations/2021_05_19_105052_create_replies_table.php
@@ -15,7 +15,7 @@ class CreateRepliesTable extends Migration
     {
         Schema::create('replies', function (Blueprint $table) {
             $table->id();
-            $table->integer('topic_id')->unsigned()->default(0)->index();
+            $table->bigInteger('topic_id')->unsigned()->default(0)->index();
             $table->bigInteger('user_id')->unsigned()->default(0)->index();
             $table->text('content');
             $table->timestamps();
